requirements.txt
GitPython==3.1.43
PyNaCl==1.5.0
click>=8.1, <8.2
dataclasses-json==0.6.7
python-dotenv>=1.0.1, <1.1
pyyaml==6.0.1
requests==2.32.3
setuptools==68.0.0
websocket-client==1.8.0
pydantic>=2.0.0, <3.0.0
pydantic-settings>=2.0.0, <3.0.0
tomli>=2.0.0, <3.0.0
tomli-w>=1.0.0, <2.0.0